Proof of concept to standarize the way we inject mocks for singletons.
This is an initial approach to a more standarized way to mock singleton
instances. That way we can increase unit test coverage testing only
isolated pieces of code.

As said this is an initial approach and the idea would be to make it
better integrated into the current testing framework.