Fixed lesson getting unassigned for teacher author
Saving course was working for admin users (users with 'manage_options' capability), but if we selected any other user as course teacher, like teacher or author, the lessons were not getting assigned to modules. Because when it tried to fetch the term for recycling, it was using slug but was not taking into consideration the slug for previous user. So could not fetch it properly.